GIAIC/20_DAYS_OF_COODING/CHATBOT_WITH_AUTHENTICATION/pyproject.toml
[project]
name = "chatbot-with-authentication"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.1",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
]
